Memory Controller MIG interface
This is currently the working version of memory controller. The part
that is left to be fixed is  the app_af_wren which is toggling whenever
addr_wren is high. This should not happen everytime. In fact it should
store the address just once and store the  next address only when write
is complete.