SPART working with the MATLAB
SPART is now working with MATLAB. The issue was with the status
register. Its value was 3 (For reading from IO address) only for one
clock cycle. Sometimes when we were lucky, cache would read the value
otherwise data was being dropped inside the hardware. Fixed it in driver
module. Now the status register will be set from 3 to 1 only when the
cache wants to read from address 800_0000, otherwise it'll be 3.